Update GenerateAccessToken function signature
The argument 'claims' in GenerateAccessToken function was taking a map with string keys and interface{} values. But all claim values in JWT should be string. Hence, updated the function argument to take a map of string keys and string values. This avoids unnecessary type assertions and can prevent potential runtime errors.