feat: add refresh token generation and tests in jwt.go & jwt_test.go
Added functionality to generate a refresh token in jwt.go. Also, included test cases for the same in jwt_test.go. The new GenerateRefreshToken function takes in an access token and secret as input and returns a signed refresh token. Functions and variables names were also refactored and code style was made consistent.